schemas, refactor

+++ rustwasm/src/grid/ds.rs
@@ -3,12 +3,12 @@ use crate::grid::core::*;
 #[derive(Default)]
 pub struct DataSource<'a> {
     pub data: &'a [SZ],
-    pub data_width: u32,
-    pub row_count: u32,
+    pub data_width: usize,
+    pub row_count: usize,
 }
 
 impl<'a> DataSource<'a> {
-    pub fn new(data: &'a [SZ], data_width: u32) -> DataSource {
+    pub fn new(data: &'a [SZ], data_width: usize) -> DataSource {
         assert!(data_width > 0);
         assert_eq!(
             data.len() as f64 % data_width as f64,
@@ -21,7 +21,7 @@ impl<'a> DataSource<'a> {
         DataSource {
             data,
             data_width,
-            row_count: (data.len() / data_width as usize) as u32,
+            row_count: (data.len() / data_width),
             ..Default::default()
         }
     }
